Add project, CD pipeline, and example function
This commit adds the bulk of the project starting with the serverless
framework, and the necessary configuration for that framework. It also
creates a CD pipeline to automatically build and deploy any commits or
merges into the main branch, to AWS Lambda. Finally, it adds an example
of a Hello World function, and an immutable response object that
contains only getter methods. While not quite a bean, this response
object does work in this case, and following its general structure may
suit this application better than a 'true bean'.